% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{partition_input}
\alias{partition_input}
\title{from rmarkdown:::partition_yaml_front_matter}
\usage{
partition_input(input_lines)
}
\value{
A list of two fields
 - front_matter: character vector of yaml metadata, typically starts and ends
  with '---'
 - body: character vector
}
\description{
from rmarkdown:::partition_yaml_front_matter
}
